requirements-latest.txt
aiohttp==3.9.5
aiosignal==1.3.1
amqp==5.2.0
anyio==4.3.0
attrs==23.2.0
backoff==2.2.1
billiard==4.2.0
blinker==1.7.0
boto3==1.26.122
botocore==1.29.122
celery==5.3.1
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
dataclasses-json==0.6.5
distro==1.9.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.0.3
frozenlist==1.4.1
greenlet==3.0.3
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
idna==3.7
invoke==2.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
jsonpatch==1.33
jsonpointer==2.4
kombu==5.3.7
langchain==0.0.352
langchain-community==0.0.19
langchain-core==0.2.26
langchain-openai==0.1.20
langfuse==1.0.18
langsmith==0.1.95
MarkupSafe==2.1.3
marshmallow==3.21.2
multidict==6.0.5
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.37.1
orjson==3.10.6
packaging==23.2
pinecone-client==3.2.2
prompt-toolkit==3.0.43
pydantic==1.10.15
pypdf==3.15.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
pytz==2023.4
PyYAML==6.0.1
redis==5.0.0
regex==2024.4.28
requests==2.31.0
s3transfer==0.6.2
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.30
tenacity==8.2.3
tiktoken==0.7.0
tqdm==4.66.4
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
uuid==1.30
vine==5.1.0
watchdog==3.0.0
wcwidth==0.2.13
Werkzeug==3.0.1
yarl==1.9.4
